requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-groq>=0.1.8

groq>=0.9.0

flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0

-e .



# langchain==0.3.26
# langchain-core==0.3.26
# langchain-community==0.3.26
# langchain-openai==0.3.26
# langchain-pinecone==0.2.8

# flask==3.1.1
# sentence-transformers==4.1.0
# pypdf==5.6.1
# python-dotenv==1.1.0

# -e .
